chore: re-organise transitions into directory
