+++ src/fmgc/src/guidance/lnav/transitions/index.ts
@@ -1,0 +1,5 @@
+import { Type1Transition } from "./Type1";
+
+export {
+    Type1Transition,
+}
